pyproject.toml
[tool.poetry]
name = "yolo-bicycle"
version = "0.1.0"
description = ""
authors = ["Juliane Coimbra <juliane.coimbra@hp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gdown = "^5.1.0"
numpy = "^1.26.4"
opencv-python-headless = "^4.9.0.80"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
opencv-contrib-python = "^4.9.0.80"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
